+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/ModelUtils.java
@@ -258,6 +258,7 @@ public class ModelUtils {
             model.templatePodSchedulerName = pod.getSchedulerName();
             model.templatePodHostAliases = pod.getHostAliases();
             model.templatePodTopologySpreadConstraints = pod.getTopologySpreadConstraints();
+            model.templatePodEnableServiceLinks = pod.getEnableServiceLinks();
         }
     }