+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Changed Zookeeper session timeout default value to 18 seconds for Topic and User Operators (for improved resiliency)
 * Removed requirement for replicas and partitions KafkaTopic spec making these parameters optional
 * Support to configure a custom filter for parent CR's labels propagation into subresources 
+* Allow disabling service links (environment variables describing Kubernetes services) in Pod template
 * Update Kaniko executor to 1.6.0
 
 ### Changes, deprecations and removals